Fixes to zoom
closes #171
