+++ app/src/main/java/com/enaboapps/switchify/service/gestures/GestureManager.kt
@@ -5,6 +5,7 @@ import android.accessibilityservice.GestureDescription
 import android.graphics.PointF
 import com.enaboapps.switchify.service.SwitchifyAccessibilityService
 import com.enaboapps.switchify.service.cursor.CursorPoint
+import com.enaboapps.switchify.service.gestures.utils.GestureUtils.getInBoundsCoordinate
 import com.enaboapps.switchify.service.gestures.visuals.GestureDrawing
 import com.enaboapps.switchify.service.utils.ScreenUtils
 import com.enaboapps.switchify.service.window.ServiceMessageHUD
@@ -36,12 +37,6 @@ class GestureManager {
 
         // This is the duration of the drag
         const val DRAG_DURATION = 1500L
-
-        // This is the duration of the zoom
-        const val ZOOM_DURATION = 2000L
-
-        // This is the amount of zoom
-        const val ZOOM_AMOUNT = 300f
     }
 
 
@@ -365,87 +360,10 @@ class GestureManager {
         return isDragging
     }
 
-    enum class ZoomAction {
-        ZOOM_IN, ZOOM_OUT
-    }
-
-    // Function to zoom in
-    fun performZoomAction(zoomAction: ZoomAction) {
-        val centerPoint =
-            CursorPoint.instance.point // Assuming this is a singleton with a current point
-        if (centerPoint != null) {
-            // Initialize paths for the two fingers
-            val path1 = android.graphics.Path()
-            val path2 = android.graphics.Path()
-
-            // Figure out the zoom point
-            val leftZoomPoint = getInBoundsCoordinate(
-                ScreenUtils.getWidth(accessibilityService!!),
-                centerPoint.x - ZOOM_AMOUNT
-            )
-            val rightZoomPoint = getInBoundsCoordinate(
-                ScreenUtils.getWidth(accessibilityService!!),
-                centerPoint.x + ZOOM_AMOUNT
-            )
-
-            when (zoomAction) {
-                ZoomAction.ZOOM_IN -> {
-                    // Setup paths for zooming in (fingers moving apart)
-                    path1.moveTo(centerPoint.x, centerPoint.y)
-                    path1.lineTo(leftZoomPoint, centerPoint.y)
-
-                    path2.moveTo(centerPoint.x, centerPoint.y)
-                    path2.lineTo(rightZoomPoint, centerPoint.y)
-                }
-
-                ZoomAction.ZOOM_OUT -> {
-                    // Setup paths for zooming out (fingers moving together)
-                    path1.moveTo(leftZoomPoint, centerPoint.y)
-                    path1.lineTo(centerPoint.x, centerPoint.y)
-
-                    path2.moveTo(rightZoomPoint, centerPoint.y)
-                    path2.lineTo(centerPoint.x, centerPoint.y)
-                }
-            }
-
-            // Create gesture description with the paths
-            val gestureBuilder = GestureDescription.Builder()
-            val stroke1 = GestureDescription.StrokeDescription(path1, 0, ZOOM_DURATION)
-            val stroke2 = GestureDescription.StrokeDescription(path2, 0, ZOOM_DURATION)
-            gestureBuilder.addStroke(stroke1).addStroke(stroke2)
-
-            val gestureDescription = gestureBuilder.build()
-
-            // Dispatch the gesture
-            try {
-                accessibilityService?.dispatchGesture(
-                    gestureDescription,
-                    object : AccessibilityService.GestureResultCallback() {
-                        override fun onCompleted(gestureDescription: GestureDescription?) {
-                            super.onCompleted(gestureDescription)
-                            // Log.d(TAG, "onCompleted")
-                        }
-                    },
-                    null
-                )
-            } catch (e: Exception) {
-                // Log.e(TAG, "onZoom: ", e)
-            }
-        } else {
-            // Handle the case where centerPoint is null
-        }
-    }
-
-    // Helper function to figure out if the gesture is going to be out of bounds
-    // Takes in two floats, the width or height of the screen, and the target coordinate
-    // Returns the target coordinate if it's within bounds, or the closest bound if it's out of bounds
-    private fun getInBoundsCoordinate(widthOrHeight: Int, target: Float): Float {
-        return if (target <= 0) {
-            0f
-        } else if (target >= widthOrHeight) {
-            widthOrHeight.toFloat()
-        } else {
-            target
+    // Function to perform a zoom
+    fun performZoomAction(zoomAction: ZoomGesturePerformer.ZoomAction) {
+        accessibilityService?.let {
+            ZoomGesturePerformer.performZoomAction(zoomAction, it)
         }
     }
