+++ app/src/main/java/com/enaboapps/switchify/service/menu/menus/gestures/ZoomGesturesMenu.kt
@@ -2,6 +2,7 @@ package com.enaboapps.switchify.service.menu.menus.gestures
 
 import com.enaboapps.switchify.service.SwitchifyAccessibilityService
 import com.enaboapps.switchify.service.gestures.GestureManager
+import com.enaboapps.switchify.service.gestures.ZoomGesturePerformer
 import com.enaboapps.switchify.service.menu.MenuItem
 import com.enaboapps.switchify.service.menu.menus.BaseMenu
 
@@ -12,11 +13,11 @@ class ZoomGesturesMenu(accessibilityService: SwitchifyAccessibilityService) :
             return listOf(
                 MenuItem("Zoom In") {
                     GestureManager.getInstance()
-                        .performZoomAction(GestureManager.ZoomAction.ZOOM_IN)
+                        .performZoomAction(ZoomGesturePerformer.ZoomAction.ZOOM_IN)
                 },
                 MenuItem("Zoom Out") {
                     GestureManager.getInstance()
-                        .performZoomAction(GestureManager.ZoomAction.ZOOM_OUT)
+                        .performZoomAction(ZoomGesturePerformer.ZoomAction.ZOOM_OUT)
                 }
             )
         }
